Fixed queNotification helper function logic
