+++ app/src/main/java/com/example/plannet/ui/entranthome/EventDetailsFragment.java
@@ -16,6 +16,7 @@ import androidx.fragment.app.Fragment;
 
 import com.example.plannet.Entrant.EntrantDBConnector;
 import com.example.plannet.Entrant.EntrantProfile;
+import com.example.plannet.Notification.EntrantNotifications;
 import com.example.plannet.R;
 import com.example.plannet.ui.entrantnotifications.EntrantNotificationsFragment;
 
@@ -121,8 +122,8 @@ public class EventDetailsFragment extends Fragment {
                                 dbConnector.updateWaitlist(userID, "pending", eventID, eventData,
                                         waitlistSuccess -> {
                                             Toast.makeText(getContext(), "Successfully added to pending waitlist (both user and event).", Toast.LENGTH_SHORT).show();
-                                            EntrantNotificationsFragment notificationsFragment = new EntrantNotificationsFragment();
-                                            notificationsFragment.queueNotification(userID, "You have been added to the pending waitlist for the event: " + title.getText().toString());
+                                            EntrantNotifications notifications = new EntrantNotifications();
+                                            notifications.queueNotification(userID, "You have been added to the pending waitlist for " + title.getText().toString());
                                         },
                                         waitlistFailure -> {
                                             Toast.makeText(getContext(), "Failed to update user's pending waitlist.", Toast.LENGTH_SHORT).show();
