+++ app/src/main/java/com/example/plannet/ui/entrantnotifications/EntrantNotificationsFragment.java
@@ -72,19 +72,4 @@ public class EntrantNotificationsFragment extends Fragment {
         return view;
     }
 
-    /**
-     * Helper method to queue a notification for the user in Firebase.
-     */
-    public void queueNotification(String userID, String message) {
-        Map<String, Object> notification = new HashMap<>();
-        // Add ID and message
-        notification.put("userIDs", userID);
-        notification.put("message", message);
-
-        //Update firebase with new information
-        firebaseDB.collection("notifications")
-                .add(notification)
-                .addOnSuccessListener(documentReference -> Log.d("NotificationHelper", "Notification queued: " + message))
-                .addOnFailureListener(e -> Log.e("NotificationHelper", "Error queuing notification", e));
-    }
 }
